chore: add refactor ideas
