+++ apps/bot/package.json
@@ -10,6 +10,7 @@
     "dotenv": "^16.4.5",
     "ethers": "^6.13.4",
     "graphql-request": "^7.1.2",
+    "tiny-invariant": "^1.3.3",
     "ts-node": "^10.9.2",
     "typescript": "^5.6.3",
     "viem": "^2.21.40"
